[Refactor] ApiService > Service-Impl 구조로 리팩토링
